% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2.R
\name{hist2}
\alias{hist2}
\title{Histogram with density and rug}
\usage{
hist2(x, bw = "SJ", ...)
}
\arguments{
\item{x}{numeric vector}

\item{bw}{bandwidth setting for \code{density}. Default is "SJ". See \code{?bw.SJ} for details and other options. Can also be a numeric value specifying the smoothing bandwidth to be used}

\item{...}{additional arguments passed to \code{hist}}
}
\value{
histogram with density and rug.
}
\description{
True histogram with overlayed density and rug. Based on code from section 8.2 of an Introduction to R.
}
\examples{
hist2(faithful$eruptions)
hist2(faithful$eruptions, breaks = 30, bw = 0.1)
}
