% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_conf_int.R
\name{mean_conf_int}
\alias{mean_conf_int}
\title{Confidence Intervals for Means}
\usage{
mean_conf_int(formula, data, fun = Hmisc::smean.cl.normal, ...)
}
\arguments{
\item{formula}{a two-sided formula of the form \code{variable ~ factor.1 + factor.2 + ...}}

\item{data}{a required data frame within which to find the variable and factor(s).}

\item{fun}{a function for calculating CIs. Default is \code{Hmisc::smean.cl.normal}}

\item{...}{arguments to pass to the function given in the fun argument, such as \code{conf.int = 0.90}}
}
\value{
A data frame with columns for means, lower, upper, and grouping variables.
}
\description{
Provides a formula interface for calculating means and associated confidence intervals within groups. Intended to return a data frame suitable for plotting or reporting.
}
\examples{
mean_conf_int(mpg ~ cyl, data = mtcars)
mean_conf_int(mpg ~ am, data = mtcars)
mean_conf_int(mpg ~ cyl + am, data = mtcars)
mean_conf_int(mpg ~ cyl + am, data = mtcars, conf.int=.90)
mean_conf_int(mpg ~ cyl + am, data = mtcars,
              fun = Hmisc::smean.cl.boot, conf.int=.90)

}
