% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_poisson_check.R
\name{plot_poisson_check}
\alias{plot_poisson_check}
\title{Poisson Model check}
\usage{
plot_poisson_check(model, dispersion = FALSE)
}
\arguments{
\item{model}{A poisson model object fitted with \code{glm}}

\item{dispersion}{A logical (default FALSE); return estimated dispersion parameter?}
}
\value{
A plot of estimated variance versus fitted mean for a Poisson count model along with \code{abline(0,1)}. The estimated variance and fitted mean should have an approximate one-to-one relationship if the Poisson assumption is satisfied. (ie, they should plot closely to the diagonal line.) If dispersion = TRUE (the default), the estimated dispersion parameter is also returned, which can be used in the call to \code{summary} if necessary.
}
\description{
Check if mean and variance of Poisson count model are roughly equivalent as described in Faraway (2006, p. 59).
}
\examples{
\dontrun{
library(faraway)
data(gala)
gala <- gala[,-2]
modp <- glm(Species ~ ., data = gala, family = poisson)
plot_poisson_check(modp)
plot_poisson_check(modp, dispersion = TRUE)
dp <- plot_poisson_check(modp)
summary(modp, dispersion = dp)
}
}
\references{
Julian Faraway. (2016) \emph{Extending the Linear Model with R}. Chapman & Hall.
}
