% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred_ellipse.R
\name{plot_pred_ellipse}
\alias{plot_pred_ellipse}
\title{Multivariate Multiple Regression model confidence ellipse}
\usage{
plot_pred_ellipse(mod, newdata, level = 0.95, ggplot = TRUE)
}
\arguments{
\item{mod}{A Multivariate Multiple Regression model fitted with \code{lm}}

\item{newdata}{works the same as the newdata argument for predict}

\item{level}{confidence level between 0 and 1. The default is 0.95}

\item{ggplot}{logical. Create plot using ggplot2? Default is TRUE}
}
\value{
plot of confidence ellipse for two response variables from a Multivariate Multiple Regression model
}
\description{
Plot ellipse to visualize the uncertainty in predictions for two responses from a Multivariate Multiple Regression model
}
\examples{
\dontrun{
# From: https://data.library.virginia.edu/getting-started-with-multivariate-multiple-regression/
ami_data <- read.table("http://static.lib.virginia.edu/statlab/materials/data/ami_data.DAT")
names(ami_data) <- c("TOT","AMI","GEN","AMT","PR","DIAP","QRS")
mlm1 <- lm(cbind(TOT, AMI) ~ GEN + AMT + PR + DIAP + QRS, data = ami_data)
mlm2 <- update(mlm1, . ~ . - PR - DIAP - QRS)
plot_pred_ellipse(mod = mlm2, newdata = data.frame(GEN = 1, AMT = 1200))}
}
\references{
Johnson, R and Wichern, D (2007). \emph{Applied Multivariate Statistical Analysis, Sixth Edition.} Prentice-Hall.
}
