% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_polr.R
\name{plot_sim_polr}
\alias{plot_sim_polr}
\title{Simulate and plot data from a polr model}
\usage{
plot_sim_polr(mod, nsim = 50, alpha = 0.15, width = 0.1)
}
\arguments{
\item{mod}{a \code{polr} model; weighted fits are not supported}

\item{nsim}{Number of simulations for \code{simulate} function; default is 50}

\item{alpha}{transparency setting for simulated data points; default is 0.25. Ranges from 0 to 1.}

\item{width}{the jitter width; default is 0.1}
}
\value{
A plot of simulated proportions along with the observed proportions of the ordered response variable from the \code{polr} model.
}
\description{
Simulate and plot data from a polr model
}
\examples{
\dontrun{
library(carData)
m <- MASS::polr(poverty ~ gender + religion + degree, data=WVS)
plot_sim_polr(m)
}
}
