% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_mod_check.R
\name{poisson_mod_check}
\alias{poisson_mod_check}
\title{Poisson Model check}
\usage{
poisson_mod_check(model, dispersion = TRUE)
}
\arguments{
\item{model}{A poisson model object fitted with `glm()`}

\item{dispersion}{A logical (default TRUE); return estimated dispersion parameter?}
}
\value{
A plot of estimated variance versus mean for a fitted poisson count model. If dispersion = TRUE (the default), the estimated dispersion parameter is also returned.
}
\description{
check if mean and variance of poisson count model are roughly equivalent;
Faraway (2006) p. 59
}
\examples{
\dontrun{
library(faraway)
data(gala)
gala <- gala[,-2]
modp <- glm(Species ~ ., data = gala, family = poisson)
poisson_mod_check(modp)
poisson_mod_check(modp, dispersion = FALSE)
dp <- poisson_mod_check(modp)
summary(modp, dispersion = dp)
}
}
\references{
Julian Faraway. (2016) \emph{Extending the Linear Model with R}. Chapman & Hall.
}
